frontend-app-api: fix circular dependency
